pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dude"
version = "0.1.0"
description = "AI coding agent with hierarchical delegation and token optimization"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Dude Developer"}
]
keywords = ["ai", "agent", "coding", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "litellm>=1.0.0",
    "google-adk>=0.1.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "fastapi[standard]>=0.104.0",
    "sqlmodel>=0.0.8",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "mcp-python-interpreter>=1.2.3",
    "multilspy>=0.0.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
docker = [
    "docker>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/dude/dude"
Repository = "https://github.com/dude/dude"
Issues = "https://github.com/dude/dude/issues"

[project.scripts]
dude = "dude.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py314"
select = ["E", "F", "I", "N", "UP", "ANN", "ASYNC"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
